ww final change requests started
OVERALL
- We are still having some issues with the legibility of the font....we
like the gray pebbled background but it's making the font a bit
ghosty....can you pick a product or one section (for a test for us to
review) and show us another option either in terms of it being bolded
or slightly heavier???  We like the font choice, so don't really want
to change that if we can avoid it.
 
HOME PAGE
- Change the Bureau of Labor statistic line in the middle to (and keep
it in quotes): "It's not just about improving traction underfoot, it's
about creating a safe work environment"
- The 3 "View More" links below the above comment should be fully
clickable anywhere on the image, not just the "View More"
- For the 3 contact option links at the bottom, please put arrow in
front of them pointing to them so that they are set off a bit more
- For the Boot icon at the lower right, can you color the "W's" white?
 
ALL PRODUCTS
- Filter: Please re-order the categories as follows:  Traction/ Walking
Condition/ Slip Hazard/ Requirements / Features / Fit
- Filter: "Extreme Ice" needs to be moved from Walking Condition to
Slip Hazard (and should be the second option below "ice and snow")
- Filter: Rename "Outdoor Only" to "Outdoor"
- Filter: Add the Filter button to the bottom above the Reset
- Please change all mentions of "Econ" in product names to "LITE"
(including the Grippers for Strippers Econ)

- GRIPS-Lite (JD3612) -- get rid of the mention of the strap kit in the
size chart  (NOTE: We have not addressed any of the product
accessories....it's not mission critical and we can handle later in a
simple fashion to be determined)

CONTACT PAGE
- This needs to be an actual page and not a contact form.  It can have
all of the contact options that we include at the bottom of the other
pages (and should), but it needs to be a page itself with a very clear
call out of 1. Our Phone Number  2. Our Email  3. Our Fax....basically,
the 3 main ways to reach