Updates
NEW DB
X - For the 3 contact option links at the bottom, please put arrow in
front of them pointing to them so that they are set off a bit more
X - For the Boot icon at the lower right, can you color the "W's" white?
X - The "Speak to an Expert" option on the product pages is bringing up
the same form as "Request a Quote", should be different
X - G4S Econ is missing the main image
X - Product Names on Product Overview Page:  Can you make each name
occupy 2 lines so that the actual name is on line 1 and the descriptive
tag is on line 2 (it's a little inconsistent right now and seems like
there is not any space limitations on length, ie. even GRIPPERS for
STRIPPERS should fit on one line)
X - Please add the "TM's" and "Circle R's" next to the product names as
I've indicated previously
X - For all Size Charts, the model numbers must be followed by the size
(ie. JD3500-XS, JD3500-S., etc.)
X - For all Size Charts...it seems like you created a basic chart and
carried it across all items...that is fine, but for sizes that don't
exist (such as Spare Spike XXXL), those rows either need to be
eliminated completed or it should have an "NA" under those sizes.
X - For Altragrips-Lite and Altragrips HP's and LP's, the HP and LP
should be spelled out on the individual product page for those 4 items
(ie.  High Profile, Low Profile).  It can remain HP and LP on the
overview page, but not in the individual page
X - Altragrips-Lite LP model number should be JD3615 (not JD36159) in
the size chart
X - GRIPS-Lite (JD3612) -- get rid of the mention of the strap kit in
the size chart  (NOTE: We have not addressed any of the product
accessories....it's not mission critical and we can handle later in a
simple fashion to be determined)
X - Heel Grips (JD250) is blank in the size chart
X - ICEGRIPS Rotor (JD4499) is blank in the size chart
X - Grippers for Strippers (JD1226) is blank in the size chart
X - GRIPS Over the sock has numeric sizes (7-13).  NOTE:  If you need
to just build a static graphic for this chart that is fine
X - ICEGRIPS Boots (JD5272/JD6272) is blank
X - The category "Slip Hazard" is missing....please make it the first
option when you add it back
X - The Filter button is missing from the top (I see you added it at
the bottom, but we still want it at the top too, but not the Reset
button (which should be bottom only))
X - Filter: "Extreme Ice" needs to be moved from Walking Condition to
Slip Hazard (and should be the second option below "ice and snow")
X - Filter: Rename "Outdoor Only" to "Outdoor"